cli/azd/resources/ai-python/requirements.txt
promptflow
promptflow-tools
azure-ai-ml
azure-ai-resources
nbconvert